# lms-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/sections": {
            "post": {
                "operationId": "SectionsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSectionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Section created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create a new course section",
                "tags": [
                    "Sections"
                ]
            }
        },
        "/sections/reorder": {
            "post": {
                "operationId": "SectionsController_reorder",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReorderSectionsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sections reordered successfully"
                    },
                    "400": {
                        "description": "Invalid section IDs supplied"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Reorder sections within a course",
                "tags": [
                    "Sections"
                ]
            }
        },
        "/sections/course/{courseId}": {
            "get": {
                "operationId": "SectionsController_findAllByCourse",
                "parameters": [
                    {
                        "name": "courseId",
                        "required": true,
                        "in": "path",
                        "description": "Course ID or Product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sections fetched successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get all sections for a course (with sessions, videos, and resources)",
                "tags": [
                    "Sections"
                ]
            }
        },
        "/sections/{id}": {
            "get": {
                "operationId": "SectionsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Section UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section fetched successfully"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get a section by ID",
                "tags": [
                    "Sections"
                ]
            },
            "patch": {
                "operationId": "SectionsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Section UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSectionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section updated successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update a section",
                "tags": [
                    "Sections"
                ]
            },
            "delete": {
                "operationId": "SectionsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Section UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete a section",
                "tags": [
                    "Sections"
                ]
            }
        },
...
        "/sessions/section/{sectionId}": {
            "get": {
                "operationId": "SessionsController_findAllBySection",
                "parameters": [
                    {
                        "name": "sectionId",
                        "required": true,
                        "in": "path",
                        "description": "Section UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions fetched successfully"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get all sessions for a section with video and resources",
                "tags": [
                    "Sessions"
                ]
            }
        },
...
        "/videos": {
            "post": {
                "operationId": "VideosController_attach",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttachVideoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video attached successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session not found"
                    },
                    "409": {
                        "description": "Video already attached to session or Vimeo ID collision"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Attach a video to a session (accepts Vimeo ID or generates dummy video)",
                "tags": [
                    "Videos"
                ]
            }
        },
        "/videos/{sessionId}": {
            "get": {
                "operationId": "VideosController_findBySessionId",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "path",
                        "description": "Session UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video fetched successfully"
                    },
                    "404": {
                        "description": "Session or video not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get video attached to a session",
                "tags": [
                    "Videos"
                ]
            },
            "patch": {
                "operationId": "VideosController_update",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "path",
                        "description": "Session UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVideoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video updated successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session or video not found"
                    },
                    "409": {
                        "description": "Vimeo ID collision"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update video metadata for a session",
                "tags": [
                    "Videos"
                ]
            },
            "delete": {
                "operationId": "VideosController_remove",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "path",
                        "description": "Session UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video detached successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session or video not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Detach / delete video from a session",
                "tags": [
                    "Videos"
                ]
            }
        },
...
    },
    "info": {
        "title": "VirtuaLMS API",
        "description": "The VirtuaLMS API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateSectionDto": {
                "type": "object",
                "properties": {
                    "courseId": {
                        "type": "string",
                        "example": "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1234",
                        "description": "Course ID or Product ID to attach this section to"
                    },
                    "title": {
                        "type": "string",
                        "example": "Module 1: Getting Started",
                        "description": "Section title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Introduction, environment setup and prerequisites",
                        "description": "Section description"
                    },
                    "position": {
                        "type": "number",
                        "example": 0,
                        "description": "Position index within the course (auto-calculated as next position if omitted)"
                    }
                },
                "required": [
                    "courseId",
                    "title"
                ]
            },
            "ReorderSectionsDto": {
                "type": "object",
                "properties": {
                    "courseId": {
                        "type": "string",
                        "example": "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1234",
                        "description": "Course ID or Product ID"
                    },
                    "sectionIds": {
                        "example": [
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1111",
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d2222",
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d3333"
                        ],
                        "description": "Array of Section IDs in the desired order (index will become the position)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "courseId",
                    "sectionIds"
                ]
            },
            "UpdateSectionDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Module 1: Advanced Fundamentals",
                        "description": "Updated section title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated section description and detailed overview",
                        "description": "Updated section description"
                    },
                    "position": {
                        "type": "number",
                        "example": 1,
                        "description": "Position index within the course"
                    }
                }
            },
...
            "AttachVideoDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "example": "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1234",
                        "description": "Session UUID to attach this video to"
                    },
                    "vimeoVideoId": {
                        "type": "string",
                        "example": "892716382",
                        "description": "Vimeo Video ID. If omitted, a dummy mock Vimeo video will be generated for development/testing."
                    },
                    "title": {
                        "type": "string",
                        "example": "Course Overview and Introduction",
                        "description": "Video title (defaults to session title if omitted)"
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed walkthrough of the module curriculum",
                        "description": "Video description"
                    },
                    "duration": {
                        "type": "number",
                        "example": 360,
                        "description": "Duration in seconds (defaults to 300 if omitted in dummy mode)"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://vumbnail.com/892716382.jpg",
                        "description": "Video thumbnail URL"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "UPLOADING",
                            "PROCESSING",
                            "READY",
                            "FAILED"
                        ],
                        "default": "READY",
                        "description": "Current video status (UPLOADING, PROCESSING, READY, FAILED)"
                    }
                },
                "required": [
                    "sessionId"
                ]
            },
            "UpdateVideoDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Video Title",
                        "description": "Video title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description of the lecture video",
                        "description": "Video description"
                    },
                    "vimeoVideoId": {
                        "type": "string",
                        "example": "918273645",
                        "description": "Vimeo Video ID"
                    },
                    "duration": {
                        "type": "number",
                        "example": 420,
                        "description": "Duration in seconds"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://vumbnail.com/918273645.jpg",
                        "description": "Video thumbnail URL"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "UPLOADING",
                            "PROCESSING",
                            "READY",
                            "FAILED"
                        ],
                        "description": "Current video status"
                    }
                }
            },
...
        }
    }
}